requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
mysql-connector-python==8.2.0
protobuf==4.21.12
pyTelegramBotAPI==4.14.0
requests==2.31.0
urllib3==2.1.0
